8.brief_version2/data/requirements.txt
pandas
beautifulsoup4
mysql-connector-python
streamlit
requests